modules/knowledge-retrieval-agent/requirements.txt
google-generativeai
sentence-transformers
scikit-learn
PyMuPDF
numpy
python-dotenv
